Support build types for ingredients
In general, when `CMAKE_BUILD_TYPE` is not specified, no compiler
options related to performance or debugging are applied by CMake.

Since `cmake-cooking` is for creating developer environments, we specify
for CMake the `Debug` build-type by default.

The build type is forwarded to all ingredients, unless it is specified
manually as part of the ingredient.