Allow recalling previous arguments for convenience
Since `cmake-cooking` refreshes all CMake caches whenever it's run, it's
necessary to supply all options at one time to configure the build.

For convenience, we save the previous invocation information in a file
on disk and allow it to be recalled with `-a` so that complicated
options do not have to be repeated.