Change the structure of ingredient arguments
This change is more robust than the previous approach because CMake needs to
know the names of all options in order to effectively parse for the values of
others (because it has to differentiate a name and a value).

Rather than list all the arguments of `ExternalProject_add` (which will change),
we just use a dedicated argument list.